refactor: moves stories/updates styles to use theme
- ContributionCountOverflow
- ContributionSummary
