+++ src/features/user/MFV2/MyContributions/MyContributions.module.scss
@@ -2,7 +2,7 @@
 
 .contributionSummary {
   padding: 6px;
-  background-color: #fff;
+  background-color: $backgroundColor;
   color: #4f4f4f;
   border-radius: 4px;
   display: flex;
@@ -37,7 +37,7 @@
 
 .contributionCountOverflow {
   background-color: #007a491a;
-  color: #007a49;
+  color: $primaryDarkColor;
   font-weight: 700;
   font-size: 10px;
   border-radius: 4px;
